Token related improvements
